move test code to test root
