+++ vertx-web/src/test/java/io/vertx/ext/web/handler/SockJSHandlerTest.java
@@ -16,6 +16,7 @@
 
 package io.vertx.ext.web.handler;
 
+import io.vertx.core.Handler;
 import io.vertx.core.MultiMap;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.CaseInsensitiveHeaders;
@@ -23,8 +24,15 @@ import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.WebTestBase;
 import io.vertx.ext.web.handler.sockjs.SockJSHandler;
+import io.vertx.ext.web.handler.sockjs.SockJSHandlerOptions;
+import io.vertx.ext.web.handler.sockjs.SockJSSocket;
 import io.vertx.test.core.TestUtils;
 import org.junit.Test;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import static io.vertx.core.buffer.Buffer.buffer;
 
 /**
  * SockJS protocol tests
@@ -33,12 +41,62 @@ import org.junit.Test;
  */
 public class SockJSHandlerTest extends WebTestBase {
 
-  private static final Logger log = LoggerFactory.getLogger(SockJSHandlerTest.class);
-
   @Override
   public void setUp() throws Exception {
     super.setUp();
-    SockJSHandler.installTestApplications(router, vertx);
+
+    // These applications are required by the SockJS protocol
+    router.route("/echo/*").handler(SockJSHandler.create(vertx,
+      new SockJSHandlerOptions().setMaxBytesStreaming(4096)).socketHandler(sock -> sock.handler(sock::write)));
+
+    router.route("/close/*").handler(SockJSHandler.create(vertx,
+      new SockJSHandlerOptions().setMaxBytesStreaming(4096)).socketHandler(SockJSSocket::close));
+
+    router.route("/disabled_websocket_echo/*").handler(SockJSHandler.create(vertx, new SockJSHandlerOptions()
+      .setMaxBytesStreaming(4096).addDisabledTransport("WEBSOCKET")).socketHandler(sock -> sock.handler(sock::write)));
+
+    router.route("/ticker/*").handler(SockJSHandler.create(vertx,
+      new SockJSHandlerOptions().setMaxBytesStreaming(4096)).socketHandler(sock -> {
+      long timerID = vertx.setPeriodic(1000, tid -> sock.write(buffer("tick!")));
+      sock.endHandler(v -> vertx.cancelTimer(timerID));
+    }));
+
+    router.route("/amplify/*").handler(SockJSHandler.create(vertx,
+      new SockJSHandlerOptions().setMaxBytesStreaming(4096)).socketHandler(sock -> {
+      sock.handler(data -> {
+        String str = data.toString();
+        int n = Integer.valueOf(str);
+        if (n < 0 || n > 19) {
+          n = 1;
+        }
+        int num = (int) Math.pow(2, n);
+        Buffer buff = buffer(num);
+        for (int i = 0; i < num; i++) {
+          buff.appendByte((byte) 'x');
+        }
+        sock.write(buff);
+      });
+    }));
+
+    router.route("/broadcast/*").handler(SockJSHandler.create(vertx,
+      new SockJSHandlerOptions().setMaxBytesStreaming(4096)).socketHandler(new Handler<SockJSSocket>() {
+      Set<String> connections = new HashSet<>();
+
+      public void handle(SockJSSocket sock) {
+        connections.add(sock.writeHandlerID());
+        sock.handler(buffer -> {
+          for (String actorID : connections) {
+            vertx.eventBus().publish(actorID, buffer);
+          }
+        });
+        sock.endHandler(v -> {
+          connections.remove(sock.writeHandlerID());
+        });
+      }
+    }));
+
+    router.route("/cookie_needed_echo/*").handler(SockJSHandler.create(vertx, new SockJSHandlerOptions().
+      setMaxBytesStreaming(4096).setInsertJSESSIONID(true)).socketHandler(sock -> sock.handler(sock::write)));
   }
 
   @Test
